Added link between picking a table and making an order
