+++ app/src/main/java/se/grupp1/antonsskafferi/PickTableFragment.java
@@ -1,0 +1,33 @@
+package se.grupp1.antonsskafferi;
+
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import androidx.annotation.NonNull;
+import androidx.fragment.app.Fragment;
+import androidx.navigation.NavController;
+import androidx.navigation.Navigation;
+
+public class PickTableFragment extends Fragment {
+
+    @Override
+    public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState)
+    {
+        View root = inflater.inflate(R.layout.fragment_pick_table, container, false);
+
+        root.findViewById(R.id.tableId1).setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                NavController navController = Navigation.findNavController(getView());
+                navController.navigate(R.id.navigation_new_order);
+            }
+        });
+
+        //Preference dinnerPreference = findPreference("editDinnerPreference");
+
+
+        return root;
+    }
+}
